+++ auth/auth-router.js
@@ -1,48 +1,48 @@
-const router = require("express").Router();
-const bcrypt = require("bcryptjs");
-const jwt = require("jsonwebtoken");
+// const router = require("express").Router();
+// const bcrypt = require("bcryptjs");
+// const jwt = require("jsonwebtoken");
 
-const Users = require("../users/users-model.js");
+// const Users = require("../users/users-model.js");
 
-router.post("/login", (req, res) => {
-  let { username, password } = req.body;
+// router.post("/login", (req, res) => {
+//   let { username, password } = req.body;
 
-  Users.findBy({ username })
-    .first()
-    .then(user => {
-      if (user && bcrypt.compareSync(password, user.password)) {
-        const token = generateToken(user);
-        res.status(200).json({
-          message: `Welcome ${user.username}!`,
-          token: token
-        });
-      } else {
-        res.status(401).json({ message: "Invalid Credentials" });
-      }
-    })
-    .catch(error => {
-      res.status(500).json(error);
-    });
-});
+//   Users.findBy({ username })
+//     .first()
+//     .then(user => {
+//       if (user && bcrypt.compareSync(password, user.password)) {
+//         const token = generateToken(user);
+//         res.status(200).json({
+//           message: `Welcome ${user.username}!`,
+//           token: token
+//         });
+//       } else {
+//         res.status(401).json({ message: "Invalid Credentials" });
+//       }
+//     })
+//     .catch(error => {
+//       res.status(500).json(error);
+//     });
+// });
 
-function generateToken(user) {
-  const payload = {
-    subject: user.id,
-    username: user.username,
-    department: user.department
-  };
-  const options = {
-    expiresIn: "1d"
-  };
+// function generateToken(user) {
+//   const payload = {
+//     subject: user.id,
+//     username: user.username,
+//     department: user.department
+//   };
+//   const options = {
+//     expiresIn: "1d"
+//   };
 
-  const result = jwt.sign(
-    payload,
-    // process.env.NODE_ENV === 'development' ? 'devsecret' : process.env.SECRET,
-    "THIS IS THE BIT THAT SHOULD BE IN AN ENV FILE SOMEWHERE",
-    options
-  );
+//   const result = jwt.sign(
+//     payload,
+//     // process.env.NODE_ENV === 'development' ? 'devsecret' : process.env.SECRET,
+//     "THIS IS THE BIT THAT SHOULD BE IN AN ENV FILE SOMEWHERE",
+//     options
+//   );
 
-  return result;
-}
+//   return result;
+// }
 
-module.exports = router;
+// module.exports = router;
